procmap: tidy Task 0b — drop duplicate comment, move ELF helper
- attachBuildIDs's inline comment duplicated its function doc; removed.
- writeELFWithBuildID (~80 LOC ELF byte layout) moved out of
  procmap_test.go into elfhelpers_test.go so the main test file
  stays focused on behavior assertions and the helper is reusable
  from any test in the package.
